chore(ci): multi-stage Dockerfile, docker-compose, GitHub Actions matrix
Dockerfile (multi-stage):
* Builder stage on python:3.13-slim + uv installs production deps into
  /app/.venv with --no-editable wheels (smaller layer + faster import).
* Runtime stage strips build tooling, runs as non-root `anpr` user
  (uid 1000), persists /data via volume, ships a curl /health
  healthcheck, exposes 8000.
* For GPU deploys, swap both bases to `nvidia/cuda:12.x-cudnn-runtime`
  and install the `gpu` extras group (onnxruntime-gpu).

docker-compose.yml: single-host deployment with ANPR_PLATE_HMAC_PEPPER
required at startup (KVKK/GDPR; never run with the auto-generated dev
fallback in production), named volume for the SQLite DB, and the same
healthcheck wired through compose.

.github/workflows/ci.yml: four parallel jobs gated on PR + push to main:

  - lint:   ruff check + ruff format --check (Python 3.13)
  - type:   pyright on src + tests
  - test:   pytest matrix (3.11/3.12/3.13), coverage xml on 3.13
  - docker: build image, run container, smoke-test /health + /version

All jobs use astral-sh/setup-uv@v5 with the pyproject lock cache
(uv handles its own caching keyed on pyproject.toml). The docker job
runs after lint/type/test pass to avoid wasting buildx cycles on
broken trees.

Plus a batch of ruff cleanup (config/ + auto-fixes + intentional-pattern
exemptions: PLC0415 lazy imports, PLW0603 module singletons, RUF001 Turkish
"ı" in province names).

+++ src/anpr/postprocess/dewarp.py
@@ -18,5 +18,5 @@ def dewarp(plate_crop: np.ndarray, *, target_height: int = 64) -> np.ndarray:
         return plate_crop
     if h == target_height:
         return plate_crop
-    target_w = max(1, int(round(w * target_height / h)))
+    target_w = max(1, round(w * target_height / h))
     return cv2.resize(plate_crop, (target_w, target_height), interpolation=cv2.INTER_CUBIC)
